+++ src/RcppExports.cpp
@@ -54,12 +54,24 @@ BEGIN_RCPP
     return R_NilValue;
 END_RCPP
 }
+// sqrt
+NumericVector sqrt(NumericVector& squared_distances);
+RcppExport SEXP _hclust1d_sqrt(SEXP squared_distancesSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< NumericVector& >::type squared_distances(squared_distancesSEXP);
+    rcpp_result_gen = Rcpp::wrap(sqrt(squared_distances));
+    return rcpp_result_gen;
+END_RCPP
+}
 
 static const R_CallMethodDef CallEntries[] = {
     {"_hclust1d_dedistance", (DL_FUNC) &_hclust1d_dedistance, 2},
     {"_hclust1d_hclust1d_heapbased", (DL_FUNC) &_hclust1d_hclust1d_heapbased, 2},
     {"_hclust1d_hclust1d_single", (DL_FUNC) &_hclust1d_hclust1d_single, 1},
     {"_hclust1d_heap_experiments", (DL_FUNC) &_hclust1d_heap_experiments, 0},
+    {"_hclust1d_sqrt", (DL_FUNC) &_hclust1d_sqrt, 1},
     {NULL, NULL, 0}
 };