+++ tests/testthat/test-hclust1d-triple_values.R
@@ -1,65 +1,6 @@
 
 repetitions <- 1:5
 
-
-test_that("points and distance length 0 and 1 should fail", {
-  set.seed(0)
-  for (tested_method in c(supported_methods(), "single_implemented_by_heap")) {
-    expect_error(hclust1d(numeric(0), method=tested_method))
-    expect_error(hclust1d(1, method=tested_method))
-    expect_error(hclust1d(dist(numeric(0)), distance = TRUE, method=tested_method))
-    expect_error(hclust1d(dist(numeric(1)), distance = TRUE, method=tested_method))
-  }
-})
-
-test_that("non-numerical points should fail", {
-  for (tested_method in c(supported_methods(), "single_implemented_by_heap")) {
-    expect_error(hclust1d("x", method=tested_method))
-  }
-})
-
-test_that("distance not logical or not scalar should fail", {
-  for (tested_method in c(supported_methods(), "single_implemented_by_heap")) {
-    expect_error(hclust1d(c(1, 2, 3), distance="yes", method=tested_method))
-    expect_error(hclust1d(c(1, 2, 3), distance=c(TRUE, TRUE), method=tested_method))
-  }
-})
-
-test_that("squared not logical or not scalar should fail", {
-  for (tested_method in c(supported_methods(), "single_implemented_by_heap")) {
-    expect_error(hclust1d(c(1, 2, 3), distance=TRUE, squared="yes", method=tested_method))
-    expect_error(hclust1d(c(1, 2, 3), distance=TRUE, squared=c(TRUE, TRUE), method=tested_method))
-  }
-})
-
-test_that("distance matrix not S3 dist class should fail", {
-  for (tested_method in c(supported_methods(), "single_implemented_by_heap")) {
-    expect_error(hclust1d(as.matrix(dist(c(1, 2, 3))), distance=TRUE, method=tested_method))
-  }
-})
-
-
-test_that("checking distance types", {
-  for (tested_method in c(supported_methods(), "single_implemented_by_heap")) {
-    #distance method should get carried over to hclust1d result
-    for (dist_method in c("euclidean", "maximum", "manhattan", "minkowski"))
-      expect_equal(hclust1d(dist(c(1, 2, 3), method=dist_method), distance=TRUE, method=tested_method)$dist.method, dist_method)
-
-    #hclust1d should fail on those two dist_methods
-    for (dist_method in c("canberra", "binary"))
-      expect_error(hclust1d(dist(c(1, 2, 3), method=dist_method), distance=TRUE, method=tested_method))
-  }
-})
-
-test_that("should preserve names or values or indices of points", {
-  for (tested_method in c(supported_methods(), "single_implemented_by_heap")) {
-    expect_equal(hclust1d(c(one=1, two=2, three=-3), method=tested_method)$labels, c("one", "two", "three"))
-    expect_equal(hclust1d(c(1.1, 2.3, -2.2), method=tested_method)$labels, c("1.1", "2.3", "-2.2"))
-    expect_equal(hclust1d(dist(c(one=1, two=2, three=-3)), distance=TRUE, method=tested_method)$labels, c("one", "two", "three"))
-    expect_equal(hclust1d(dist(c(1.1, 2.3, -2.2)), distance=TRUE, method=tested_method)$labels, c("1", "2", "3"))
-  }
-})
-
 expect_some_equalities <- function(res_1, res_2) {
   if (res_1$method == "single_implemented_by_heap") {
     res_1$method <- "single"
@@ -74,12 +15,6 @@ expect_some_equalities <- function(res_1, res_2) {
   expect_equal(res_1$height, res_2$height)
 
   expect_s3_class(res_1, class(res_2))
-}
-
-expect_equal_merges <- function(res_1, res_2) {
-  for (i in 1:nrow(res_1$merge)) {
-    expect_setequal(res_1$merge[i, ], res_2$merge[i, ])
-  }
 }
 
 expect_all_equalities <- function(res_1d, res_1d_alt, res_1d_dist) {
@@ -116,113 +51,6 @@ expect_all_equalities <- function(res_1d, res_1d_alt, res_1d_dist) {
   }
 
 }
-
-
-range <- 2:100
-test_that("equality of results with stats::hclust, a vector without repetitions", {
-  for (tested_method in c(supported_methods(), "single_implemented_by_heap")) {
-    stats_hlust_method <- tested_method
-    if (tested_method == "single_implemented_by_heap")
-      stats_hlust_method <- "single"
-    for (len in range) {
-      for (j in repetitions) {
-        x <- rnorm(len)
-        res_1d <- hclust1d(x, method=tested_method)
-
-        squared <- FALSE
-        distance <- dist(x)
-        if (tested_method == "centroid") {
-          squared <- TRUE
-          distance <- distance^2
-        }
-
-        res_1d_dist <- hclust1d(distance, distance = TRUE, squared = squared, method=tested_method)
-        res_1d_alt <- hclust1d(-x, method=tested_method)
-        res_md <- stats::hclust(distance, method=stats_hlust_method)
-
-       # expect_all_equalities(res_1d, res_1d_alt, res_1d_dist)
-
-        expect_some_equalities(res_1d, res_md)
-        for (i in 1:nrow(res_1d$merge)) {
-          expect_setequal(res_1d$merge[i, ], res_md$merge[i, ])
-        }
-      }
-    }
-  }
-})
-
-get_order_double <- function(res) order(vapply(1:nrow(res$merge), function(i) if (min(res$merge[i, ]) < 0) min(res$merge[i, ]) else i, numeric(1)))
-
-range <- 1:80
-percent <- 0.25
-test_that("equality of results with stats::hclust, a vector with double repetitions", {
-  # with double repetitions the clusters comprising of two singletons should be the same as in stats::hclust
-  # but may be differently ordered
-  for (tested_method in c(supported_methods(), "single_implemented_by_heap")) {
-    stats_hlust_method <- tested_method
-    if (tested_method == "single_implemented_by_heap")
-      stats_hlust_method <- "single"
-    for (len in range) {
-      for (j in repetitions) {
-        x <- rnorm(len)
-        cnt_to_add <- max(1, len * percent)
-
-        new_x_indices <- c(1:length(x), 1:cnt_to_add)
-        new_x <- x[new_x_indices]
-
-        shuffle_vector <- sample(1:length(new_x), length(new_x), replace = FALSE) #random permutation of new_x
-
-        x <- new_x[shuffle_vector]
-
-        res_1d <- hclust1d(x, method=tested_method)
-
-        squared <- FALSE
-        distance <- dist(x)
-        if (tested_method == "centroid") {
-          squared <- TRUE
-          distance <- distance^2
-        }
-
-        res_1d_dist <- hclust1d(distance, distance = TRUE, squared = squared, method=tested_method)
-        res_1d_alt <- hclust1d(-x, method=tested_method)
-        res_md <- stats::hclust(distance, method=stats_hlust_method)
-
-        expect_all_equalities(res_1d, res_1d_alt, res_1d_dist)
-        expect_some_equalities(res_1d, res_md)
-
-        #because of many heights==0 merge levels, we don't know the order of singleton merge. Fix it
-        order_1d <- get_order_double(res_1d)
-        order_md <- get_order_double(res_md)
-
-      #   ALWAYS x == x[order(x)][order(order(x))] , order(order(x)) is the inverse permutation of order(x)
-
-        for (i in 1:nrow(res_1d$merge)) {  #the condition is more complex here
-          # the corresponding steps are order_1d[i] and order_md[i]
-          for (j in 1:2) {
-            if (res_1d$merge[order_1d[i], j] >= 1) {  # a cluster merged in a previous algorithm step
-
-              found <- FALSE   #seek the the step number within the hclust results for the corresponding step
-              for (k in 1:2) {
-                if (res_md$merge[order_md[i], k] >= 1) {
-                  if (order(order_1d)[res_1d$merge[order_1d[i], j]] == order(order_md)[res_md$merge[order_md[i], k]]) {
-                    found <-TRUE   # must compare the correspondence of the previous step in between methods
-                    # seek the correspondence in inverse permutation of corresponding  orders
-                  }
-                }
-              }
-
-              expect_true(found)
-            } else {  # a singleton
-
-              expect_true(res_1d$merge[order_1d[i], j] %in% res_md$merge[order_md[i], ])
-            }
-          }
-        }
-
-      }
-    }
-  }
-})
 
 get_links_triple <- function(res) {
   # finds links between 0-height singleton-only clusters and the second cluster in which a third singleton is added with height 0
@@ -270,7 +98,7 @@ test_that("equality of results with stats::hclust, a vector with triple repetiti
 
   # with more than triple repetitions it would become intractable in principle,
   # what would be the order of creation of larger clusters
-  for (tested_method in c(supported_methods(), "single_implemented_by_heap")) {
+  for (tested_method in c(supported_methods(), "single_implemented_by_heap")[-4]) {  #without a test for true_median
     stats_hlust_method <- tested_method
     if (tested_method == "single_implemented_by_heap")
       stats_hlust_method <- "single"
@@ -292,7 +120,7 @@ test_that("equality of results with stats::hclust, a vector with triple repetiti
 
         squared <- FALSE
         distance <- dist(x)
-        if (tested_method == "centroid") {
+        if (tested_method %in% c("centroid", "median")) {
           squared <- TRUE
           distance <- distance^2
         }